To avoid some display mistake
In some case, the header.datasize(6) would be uint32. Then in the
following step, (tpx-1)*header.xstep would be also uint32. then the
figures would display incorrectly.

+++ core_functions/CGLW_multi_viewer.m
@@ -446,7 +446,7 @@ for datasetpos=1:length(tdata);
     %header
     header=datasets_header(selected_datasets(datasetpos)).header;
     %tpx
-    tpx=1:1:header.datasize(6);
+    tpx=1:1:double(header.datasize(6));
     tpx=((tpx-1)*header.xstep)+header.xstart;
     for epochpos=1:length(selected_epochs);
         for chanpos=1:length(channel_idx);
